update settings and cleanup

+++ scripts/Cleanup.ps1
@@ -1,17 +1,31 @@
 ﻿. (Join-Path $PSScriptRoot "Read-Settings.ps1")
+if ("$AgentName" -ne "Hosted Agent" -and "$AgentName" -notlike "Azure Pipelines*") {
+    . (Join-Path $PSScriptRoot "Install-BcContainerHelper.ps1") -bcContainerHelperVersion $bcContainerHelperVersion
 
-$bcContainerHelperVersion = "latest"
-if ($settings.PSObject.Properties.Name -eq 'bcContainerHelperVersion' -and $settings.bcContainerHelperVersion) {
-    $bcContainerHelperVersion = $settings.bcContainerHelperVersion
-}
-Write-Host "Use bcContainerHelper Version: $bcContainerHelperVersion"
-. (Join-Path $PSScriptRoot "Install-BcContainerHelper.ps1") -bcContainerHelperVersion $bcContainerHelperVersion
-
-Remove-BcContainer -containerName $containerName
-Flush-ContainerHelperCache -KeepDays 7
-
-Remove-Module BcContainerHelper
-$path = Join-Path $ENV:Temp $containerName
-if (Test-Path $path) {
-    Remove-Item $path -Recurse -Force
+    $cleanupMutexName = "Cleanup"
+    $cleanupMutex = New-Object System.Threading.Mutex($false, $cleanupMutexName)
+    try {
+        try {
+            if (!$cleanupMutex.WaitOne(1000)) {
+                Write-Host "Waiting for other process to finish cleanup"
+                $buildMutex.WaitOne() | Out-Null
+                Write-Host "Other process completed"
+            }
+        }
+        catch [System.Threading.AbandonedMutexException] {
+           Write-Host "Other process terminated abnormally"
+        }
+    
+        Remove-BcContainer -containerName $containerName
+        Flush-ContainerHelperCache -KeepDays 2
+        
+        Remove-Module BcContainerHelper
+        $path = Join-Path $ENV:Temp $containerName
+        if (Test-Path $path) {
+            Remove-Item $path -Recurse -Force
+        }
+    }
+    finally {
+        $cleanupMutex.ReleaseMutex()
+    }
 }
